Kompleksnejsa poizvedba z 2 parametroma

+++ FastCare/src/main/java/com/test/fastcare/vao/Urgenca.java
@@ -10,19 +10,18 @@ public class Urgenca {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
 
-    private long id_urgenca;
+    private int id_urgenca;
     private String nazivUrgence;
     private String naslovUrgence;
     private String sobe;
     private String oddelki;
     private String delavci;
 
-    public long getId_urgenca() {
+    public int getId_urgenca() {
         return id_urgenca;
     }
 
-
-    public void setId_urgenca(long id_urgenca) {
+    public void setId_urgenca(int id_urgenca) {
         this.id_urgenca = id_urgenca;
     }
